+++ obj/src/auth/DefaultCredentialStoreFactory.d.ts
@@ -3,5 +3,6 @@ import { Descriptor } from '../refer/Descriptor';
 export declare class DefaultCredentialStoreFactory extends Factory {
     static readonly Descriptor: Descriptor;
     static readonly MemoryCredentialStoreDescriptor: Descriptor;
+    static readonly MemoryCredentialStoreDescriptor2: Descriptor;
     constructor();
 }