+++ obj/src/connect/DefaultDiscoveryFactory.d.ts
@@ -3,5 +3,6 @@ import { Descriptor } from '../refer/Descriptor';
 export declare class DefaultDiscoveryFactory extends Factory {
     static readonly Descriptor: Descriptor;
     static readonly MemoryDiscoveryDescriptor: Descriptor;
+    static readonly MemoryDiscoveryDescriptor2: Descriptor;
     constructor();
 }