+++ obj/src/count/DefaultCountersFactory.d.ts
@@ -3,7 +3,10 @@ import { Descriptor } from '../refer/Descriptor';
 export declare class DefaultCountersFactory extends Factory {
     static readonly Descriptor: Descriptor;
     static readonly NullCountersDescriptor: Descriptor;
+    static readonly NullCountersDescriptor2: Descriptor;
     static readonly LogCountersDescriptor: Descriptor;
+    static readonly LogCountersDescriptor2: Descriptor;
     static readonly CompositeCountersDescriptor: Descriptor;
+    static readonly CompositeCountersDescriptor2: Descriptor;
     constructor();
 }