+++ obj/src/lock/index.d.ts
@@ -1,5 +1,5 @@
 export { ILock } from './ILock';
+export { Lock } from './Lock';
 export { NullLock } from './NullLock';
 export { MemoryLock } from './MemoryLock';
-export { MemcachedLock } from './MemcachedLock';
 export { DefaultLockFactory } from './DefaultLockFactory';