+++ obj/src/lock/index.js
@@ -1,11 +1,11 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+var Lock_1 = require("./Lock");
+exports.Lock = Lock_1.Lock;
 var NullLock_1 = require("./NullLock");
 exports.NullLock = NullLock_1.NullLock;
 var MemoryLock_1 = require("./MemoryLock");
 exports.MemoryLock = MemoryLock_1.MemoryLock;
-var MemcachedLock_1 = require("./MemcachedLock");
-exports.MemcachedLock = MemcachedLock_1.MemcachedLock;
 var DefaultLockFactory_1 = require("./DefaultLockFactory");
 exports.DefaultLockFactory = DefaultLockFactory_1.DefaultLockFactory;
 //# sourceMappingURL=index.js.map