+++ src/cache/index.ts
@@ -2,5 +2,4 @@ export { CacheEntry } from './CacheEntry';
 export { ICache } from './ICache';
 export { NullCache } from './NullCache';
 export { MemoryCache } from './MemoryCache';
-export { MemcachedCache } from './MemcachedCache';
 export { DefaultCacheFactory } from './DefaultCacheFactory';