+++ src/connect/DefaultDiscoveryFactory.ts
@@ -3,11 +3,13 @@ import { Descriptor } from '../refer/Descriptor';
 import { MemoryDiscovery } from './MemoryDiscovery';
 
 export class DefaultDiscoveryFactory extends Factory {
-	public static readonly Descriptor = new Descriptor("pip-services-commons", "factory", "discovery", "default", "1.0");
-	public static readonly MemoryDiscoveryDescriptor = new Descriptor("pip-services-commons", "discovery", "memory", "*", "1.0");
+	public static readonly Descriptor = new Descriptor("pip-services", "factory", "discovery", "default", "1.0");
+	public static readonly MemoryDiscoveryDescriptor = new Descriptor("pip-services", "discovery", "memory", "*", "1.0");
+	public static readonly MemoryDiscoveryDescriptor2 = new Descriptor("pip-services-commons", "discovery", "memory", "*", "1.0");
 	
 	public constructor() {
         super();
 		this.registerAsType(DefaultDiscoveryFactory.MemoryDiscoveryDescriptor, MemoryDiscovery);
+		this.registerAsType(DefaultDiscoveryFactory.MemoryDiscoveryDescriptor2, MemoryDiscovery);
 	}	
 }